Fix linters
