+++ index.css
@@ -1,187 +1,186 @@
 body {
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-    height: 90vh;
-    margin: 0;
-    padding: 1rem 8rem;
-  }
-  
-  nav {
-    display: flex;
-    width: 100%;
-    justify-content: space-around;
-    height: 2rem;
-  
-    /* padding: 1rem 10rem; */
-    border: 5px solid #000;
-    align-items: center;
-  }
-  
-  nav h2 {
-    margin: 0;
-  }
-  
-  nav ul {
-    display: flex;
-    padding: 0;
-    gap: 2rem;
-    margin: 0;
-  }
-  
-  nav ul li {
-    list-style: none;
-    cursor: pointer;
-    font-size: 1.5rem;
-  }
-  
-  .list ul li {
-    list-style: none;
-    margin: 0;
-    font-size: 2rem;
-  }
-  
-  nav ul li:nth-child(2) {
-    border-right: 2px solid #000;
-    border-left: 2px solid #000;
-    padding: 0 2rem;
-  }
-  
-  .date {
-    text-align: end;
-    font-size: 1.5rem;
-    padding-top: 1rem;
-  }
-  
-  .main {
-    width: 100%;
-    display: flex;
-    align-items: center;
-    flex-direction: column;
-    justify-content: center;
-    height: auto;
-  }
-  
-  form h1 {
-    margin: 0;
-    font-size: 3rem;
-  }
-  
-  .main h1 {
-    font-size: 3rem;
-  }
-  
-  .main.active {
-    display: none;
-  }
-  
-  .list {
-    border: 5px solid #000;
-    width: 100%;
-    height: auto;
-  }
-  
-  .list ul {
-    display: flex;
-    align-items: center;
-    padding: 1rem 2rem;
-    margin: 0;
-    gap: 8px;
-    height: 3rem;
-  }
-  
-  .contact-section ul {
-    font-size: 1.5rem;
-  }
-  
-  .list ul:nth-child(odd) {
-    background-color: rgb(188, 188, 194);
-  }
-  
-  .list ul:nth-child(even) {
-    background-color: rgb(255, 255, 255);
-  }
-  
-  .delete {
-    margin-left: auto;
-    background-color: white;
-    box-shadow: 5px 5px 5px black;
-    cursor: pointer;
-    font-size: 1.5rem;
-  }
-  
-  .list.active {
-    display: none;
-  }
-  
-  form {
-    display: none;
-  }
-  
-  form.active {
-    height: 100%;
-    padding-top: 1.5rem;
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    gap: 5rem;
-  }
-  
-  .inputs {
-    width: 40%;
-    display: flex;
-    flex-direction: column;
-    gap: 3rem;
-  }
-  
-  form input {
-    height: 3rem;
-    border: 5px solid #000;
-    padding: 0.5rem;
-    font-size: 2rem;
-  }
-  
-  form input::placeholder {
-    font-size: 2rem;
-  }
-  
-  form button {
-    background-color: white;
-    box-shadow: 5px 5px 5px black;
-    align-self: flex-end;
-    width: 10%;
-    font-size: 1.5rem;
-  }
-  
-  footer {
-    display: flex;
-    width: 100%;
-    justify-content: space-around;
-    height: 2rem;
-  
-    /* padding: 1rem 10rem; */
-    border: 5px solid #000;
-    align-items: center;
-    margin-top: auto;
-  }
-  
-  .contact-section.active {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-  }
-  
-  .contact-section h2 {
-    font-size: 3rem;
-  }
-  
-  .contact-section p {
-    font-size: 1.5rem;
-  }
-  
-  .contact-section {
-    display: none;
-  }
-  
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+  height: 90vh;
+  margin: 0;
+  padding: 1rem 8rem;
+}
+
+nav {
+  display: flex;
+  width: 100%;
+  justify-content: space-around;
+  height: 2rem;
+
+  /* padding: 1rem 10rem; */
+  border: 5px solid #000;
+  align-items: center;
+}
+
+nav h2 {
+  margin: 0;
+}
+
+nav ul {
+  display: flex;
+  padding: 0;
+  gap: 2rem;
+  margin: 0;
+}
+
+nav ul li {
+  list-style: none;
+  cursor: pointer;
+  font-size: 1.5rem;
+}
+
+.list ul li {
+  list-style: none;
+  margin: 0;
+  font-size: 2rem;
+}
+
+nav ul li:nth-child(2) {
+  border-right: 2px solid #000;
+  border-left: 2px solid #000;
+  padding: 0 2rem;
+}
+
+.date {
+  text-align: end;
+  font-size: 1.5rem;
+  padding-top: 1rem;
+}
+
+.main {
+  width: 100%;
+  display: flex;
+  align-items: center;
+  flex-direction: column;
+  justify-content: center;
+  height: auto;
+}
+
+form h1 {
+  margin: 0;
+  font-size: 3rem;
+}
+
+.main h1 {
+  font-size: 3rem;
+}
+
+.main.active {
+  display: none;
+}
+
+.list {
+  border: 5px solid #000;
+  width: 100%;
+  height: auto;
+}
+
+.list ul {
+  display: flex;
+  align-items: center;
+  padding: 1rem 2rem;
+  margin: 0;
+  gap: 8px;
+  height: 3rem;
+}
+
+.contact-section ul {
+  font-size: 1.5rem;
+}
+
+.list ul:nth-child(odd) {
+  background-color: rgb(188, 188, 194);
+}
+
+.list ul:nth-child(even) {
+  background-color: rgb(255, 255, 255);
+}
+
+.delete {
+  margin-left: auto;
+  background-color: white;
+  box-shadow: 5px 5px 5px black;
+  cursor: pointer;
+  font-size: 1.5rem;
+}
+
+.list.active {
+  display: none;
+}
+
+form {
+  display: none;
+}
+
+form.active {
+  height: 100%;
+  padding-top: 1.5rem;
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  gap: 5rem;
+}
+
+.inputs {
+  width: 40%;
+  display: flex;
+  flex-direction: column;
+  gap: 3rem;
+}
+
+form input {
+  height: 3rem;
+  border: 5px solid #000;
+  padding: 0.5rem;
+  font-size: 2rem;
+}
+
+form input::placeholder {
+  font-size: 2rem;
+}
+
+form button {
+  background-color: white;
+  box-shadow: 5px 5px 5px black;
+  align-self: flex-end;
+  width: 10%;
+  font-size: 1.5rem;
+}
+
+footer {
+  display: flex;
+  width: 100%;
+  justify-content: space-around;
+  height: 2rem;
+
+  /* padding: 1rem 10rem; */
+  border: 5px solid #000;
+  align-items: center;
+  margin-top: auto;
+}
+
+.contact-section.active {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.contact-section h2 {
+  font-size: 3rem;
+}
+
+.contact-section p {
+  font-size: 1.5rem;
+}
+
+.contact-section {
+  display: none;
+}
