+++ index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="style.css" />
+    <link rel="stylesheet" href="index.css" />
     <title>Awesome Books</title>
   </head>
 
@@ -43,8 +43,6 @@
       <h3 class="footer-text">Copyright. All rights reserved.</h3>
       <h3>Developed by 3 cool dudes</h3>
     </footer>
-    <!-- <script src="./book.js"></script> -->
-    <script src="script.js"></script>
-    <!-- <script type="module" src="./modelJS/index.js"></script> -->
+    <script src="./index.js" type="module"></script>
   </body>
 </html>
