+++ index.js
@@ -1,98 +1,38 @@
-/* eslint-disable max-classes-per-file */
-const title = document.getElementById('title');
-const author = document.getElementById('author');
-const button = document.getElementById('button');
+import Classbook from './modules/classbook.js';
+import Storage from './modules/classstorage.js';
+import display from './modules/classdisplay.js';
+import {
+  title,
+  author,
+  button,
+  addNew,
+  form,
+  main,
+  showMain,
+  showContact,
+  contact,
+  date,
+} from './modules/variables-fullpage.js';
 
-class Book {
-  constructor(title, author) {
-    this.title = title;
-    this.author = author;
-  }
-}
-
-class Storage {
-  static getBooks() {
-    let booksArray;
-    if (localStorage.getItem('books') === null) {
-      booksArray = [];
-    } else {
-      booksArray = JSON.parse(localStorage.getItem('books'));
-    }
-    return booksArray;
-  }
-
-  static addBookStorage(book) {
-    const booksArray = Storage.getBooks();
-    booksArray.push(book);
-    localStorage.setItem('books', JSON.stringify(booksArray));
-  }
-
-  static removeBookStorage(bookTitle) {
-    let booksArray = Storage.getBooks();
-    booksArray = booksArray.filter((book) => book.title !== bookTitle);
-    localStorage.setItem('books', JSON.stringify(booksArray));
-  }
-}
-
-class Display {
-  static displayList() {
-    const booksArray = Storage.getBooks();
-    booksArray.forEach((book) => Display.addBook(book));
-  }
-
-  static addBook(book) {
-    const list = document.getElementById('list');
-    const bookCard = document.createElement('ul');
-    const title = document.createElement('li');
-    const author = document.createElement('li');
-    const removeButton = document.createElement('button');
-    removeButton.classList.add('delete');
-
-    title.innerHTML = book.title;
-    author.textContent = `by ${book.author}`;
-    removeButton.textContent = 'Remove';
-
-    list.appendChild(bookCard);
-    bookCard.append(title, author, removeButton);
-  }
-
-  static removeBook(eTarget) {
-    if (eTarget.classList.contains('delete')) {
-      eTarget.parentElement.remove();
-    }
-  }
-
-  static clearInputFields() {
-    title.value = '';
-    author.value = '';
-  }
-}
+import { DateTime } from './node_modules/luxon/src/luxon.js';
 
 button.addEventListener('click', (e) => {
   e.preventDefault();
-  const book = new Book(title.value, author.value);
-  Display.addBook(book);
+  const book = new Classbook(title.value, author.value);
+  display.addBook(book);
   Storage.addBookStorage(book);
-  Display.clearInputFields();
+  display.clearInputFields();
 });
 
 document.getElementById('list').addEventListener('click', (e) => {
-  Display.removeBook(e.target);
+  display.removeBook(e.target);
+
   Storage.removeBookStorage(e.target.parentElement.firstChild.textContent);
 });
 
-Display.displayList();
+display.displayList();
 
 // full page events
-
-const addNew = document.querySelector('.add-new');
-const form = document.querySelector('.form');
-const main = document.querySelector('.main');
-const showMain = document.querySelector('.show-main');
-const showContact = document.querySelector('.show-contact');
-const contact = document.querySelector('.contact-section');
-
-const date = document.querySelector('.date');
 
 addNew.addEventListener('click', () => {
   if (!form.classList.contains('active')) {
@@ -118,4 +58,5 @@ showContact.addEventListener('click', () => {
   }
 });
 
-date.innerHTML = Date();
+const now = DateTime.local().toLocaleString(DateTime.DATETIME_MED);
+date.innerHTML = now;
