+++ modules/variables-fullpage.js
@@ -1,0 +1,24 @@
+const title = document.getElementById('title');
+const author = document.getElementById('author');
+const button = document.getElementById('button');
+
+const addNew = document.querySelector('.add-new');
+const form = document.querySelector('.form');
+const main = document.querySelector('.main');
+const showMain = document.querySelector('.show-main');
+const showContact = document.querySelector('.show-contact');
+const contact = document.querySelector('.contact-section');
+const date = document.querySelector('.date');
+
+export {
+  title,
+  author,
+  button,
+  addNew,
+  form,
+  main,
+  showMain,
+  showContact,
+  contact,
+  date,
+};
